feat: Begin zola migration
I thought it would be cool to use Next.js instead of Gatsby, however I
quickly realized that Next.js makes it non-trivial to do things such as
reference images in markdown relative to the current directory.

Zola handles all of this and much more out of the box, which improves
the portability of posts since there's no longer a dependence on absolute
paths. This also solves the problem with internal links not being
addressed properly inside markdown files.