chore: Add next-remote-watch
This was necessary for live reloading when updating the markdown files.

Although Next.js makes it possible to use markdown for a blog, I feel
like it's better to "use the right tool for the job" and prefer Next.js
for interactive web applications that don't have a lot of static files.