Revert "feat: Remove /blog from blog posts"
It's probably easier if I keep /blog so it helps people click on "blog"
to see more blog posts, similar to how having "projects" in the URL is
useful.

Additionally, the /blog part helps make it sound like it's part of the
blog and not something I'm doing right now.