meta: Add attempt at slug page for posts
This was my attempt at using slug pages for posts. Although this
*worked*, it definitely felt like trying to use Next.js for something
it wasn't meant to do.

Although contentlayer solves some of these problems such as the lack of
reloading, it's not currently being maintained and users are expected to
run older versions of Next.js. Furthermore, even though it's possible to
get fast refresh working with next-mdx-remote, images don't appear to
work from remark-mdx-images.

Overall it seems better to prefer the app router here since the content
is static, changes in development, and not dynamic.