fix: Expand all code blocks by default
Fixes an issue where .md files no longer respect custom HTML tags inside
them. In the future it may be possible to check the length of the
longest line in the code block before deciding whether or not to add a
class that expands it.

The advantage of this is that how the markdown is displayed happens at
the app layer instead of the content layer, which shouldn't have
seemingly random <figure>s in its code.